implicit conversion double/CGFloat when compiling for 32-bit

+++ Classes/BITHockeyHelper.m
@@ -776,9 +776,9 @@ UIImage *bit_screenshot(void) {
       
       if (needsRotation) {
         if (isLandscapeLeft) {
-          CGContextConcatCTM(context, CGAffineTransformRotate(CGAffineTransformMakeTranslation( imageSize.width, 0), M_PI / 2.0));
+          CGContextConcatCTM(context, CGAffineTransformRotate(CGAffineTransformMakeTranslation( imageSize.width, 0), (CGFloat)M_PI / 2.0f));
         } else if (isLandscapeRight) {
-          CGContextConcatCTM(context, CGAffineTransformRotate(CGAffineTransformMakeTranslation( 0, imageSize.height), 3 * M_PI / 2.0));
+          CGContextConcatCTM(context, CGAffineTransformRotate(CGAffineTransformMakeTranslation( 0, imageSize.height), 3.0f * (CGFloat)M_PI / 2.0f));
         }
       } else if (isUpsideDown) {
         CGContextConcatCTM(context, CGAffineTransformRotate(CGAffineTransformMakeTranslation( imageSize.width, imageSize.height), (CGFloat)(M_PI)));
